Promote decoding to a first-class mdopt.decoding module
decoding.py and dem.py move from mdopt/examples/decoding/ to a new
mdopt/decoding/ package with a public-API __init__. The old import paths keep
working through deprecation shims that alias their sys.modules entries to the
moved modules, so every name (public or private) resolves to the same objects
while emitting a DeprecationWarning.

The experiment scripts and their plotting stay in mdopt/examples/decoding/
and now import from the new location, as do the tests and the example
notebooks (import lines rewritten in place, outputs untouched). The
gauge-seed test derived the scripts' package from decode_css.__module__,
which the split breaks; it now names mdopt.examples.decoding explicitly.

Promotion also puts both files under mypy for the first time (examples/ is
excluded); the 19 pre-existing annotation defects that surfaced are fixed:
honest two-element tuple returns, the erasure model requiring its rate, a
stringified qec.noise_model annotation that no longer resolves, narrowed
Union pipelines in decode_custom/decode_dem, and apply_constraints declaring
List[List[int]] for what every caller passes as List[List[List[int]]].
