+++ mdopt/examples/decoding/plotting/gen_hist_shor.py
@@ -26,7 +26,7 @@ import matplotlib.pyplot as plt
 from collections import Counter
 from tqdm import tqdm
 
-from mdopt.examples.decoding.decoding import decode_css, map_distribution_to_pauli
+from mdopt.decoding.decoding import decode_css, map_distribution_to_pauli
 
 mpl.rcParams.update(
     {
