+++ mdopt/examples/decoding/plotting/gen_qubit_order_bb.py
@@ -36,7 +36,7 @@ from matplotlib import colormaps
 from matplotlib.colors import Normalize
 from tqdm import tqdm
 
-from mdopt.examples.decoding.decoding import (
+from mdopt.decoding.decoding import (
     create_bb_code,
     generate_pauli_error_string,
     css_code_checks,
